requirements.txt
Flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
Pillow==10.0.0
pyppeteer==1.0.2
gunicorn==21.2.0
